Fix: Load filters as soon as they complete downloading.
Before this change, syncdown(...) would bail early when:
1. A node was downloading a filter.
2. A node completed downloading a filter and the rules were reloaded.

1. makes sense as the filters are, in a way, undefined until the ignore
file completes donwloading.

2. was an artifact of us relying on the nchildren map for quickly
determining whether an ignore file is or was being downloaded. By the
time we had determined the filter was no longer downloading, we'd
already excluded a bunch of remote nodes.

The solution to 2. is to split that logic so that we populate the
nchildren map and cut it down to size later. This gives us space to
reload the filters if necessary.

Finally, we no longer bail by returning false.